requirements.txt
pandas
kafka-python
scikit-learn
joblib 
streamlit 